requirements.txt
asgiref==3.4.1
bitly-api-py3==0.3
django-dirtyfields==1.7.0
django-docutils==0.5.1
django-extensions==3.1.3
django-randomslugfield==0.3.0
django-slugify-processor==0.11.0
Django==3.2.7
docutils==0.17.1
lxml==4.6.3
pip==21.2.4
Pygments==2.10.0
pytz==2021.1
setuptools==57.5.0
six==1.16.0
sqlparse==0.4.1
wheel==0.37.0
django-ckeditor~=6.1.0